Add files via upload
This project was for my summer semester class titled Data Analytics Programming. The course was structured to teach students the basics of Python and Pandas. The instructions for this assignment was to investigate any dataset and try to solve a problem of our choice. The problem that I tried to solve was to accurately predict whether or not a patient in the Pima Indian heritage has diabetes. The dataset (pima_diabetes.csv) was taken from the National Institute of Diabetes and Digestive and Kidney Diseases.