Add files via upload
This project was for my fall semester class titled Introduction to Data Management. The course was structured to teach students the basics of SQL. For this project we had to select a real world company, create three transaction management applications along with a data warehouse, and select a subset of the data to analyze any market trends.