Add files via upload
This project was for my fall semester class titled Text Analysis. The course was structured to teach students how to scrape and analyze real time, unstructured data. For this project, we constructed a beer recommendation system based on user defined attributes. For the recommendation system, we performed Cosine Similarity and feature-level Analyses to capture similar beers influenced by user reviews’ sentiment and rankings.