Add files via upload
This project was for my summer semester class titled Predictive Modeling. The course was structured to teach students the basics of R. The instructions for this assignment was to find an interesting prediction/classification problem and develop the best possible predictive algorithm for the variable of interest. For this assignment, we constructed a model to forecast actual house sale prices in Ames, IA.