Add files via upload
This project was for my fall semester class titled Advanced Predictive Modeling. The course was structured to teach students the advanced predictive and classification models and regressions. The instructions for this assignment was to investigate any dataset and try to solve a problem of our choice. For this assignment, we analyzed customer browsing patterns on Expedia to predict which hotel cluster they will book. The assignment was based on a past Kaggle Competition (https://www.kaggle.com/c/expedia-hotel-recommendations). We then had to do a blog write up of the entire project (https://petedavis22.medium.com/seo-search-expedia-optimization-779d728b1374).