Add files via upload
This project was for my fall semester class titled Asset Management. The course was structured to teach students asset management in a quantitative manner. The instructions for this assignment was to replicate the S&P 500 and Dow Jones Industrial Average Indices and to compare the replications with the actual index and a passive ETF and Mutual Fund.